feat(polygon): add polyline representation between points
Factor out watchState to ToolWidgetUtils
