+++ src/vtk/ToolWidgetUtils/utils.ts
@@ -1,0 +1,13 @@
+export function watchState(
+  publicAPI: any,
+  state: any,
+  callback: () => unknown
+) {
+  let subscription = state.onModified(callback);
+  const originalDelete = publicAPI.delete;
+  publicAPI.delete = () => {
+    subscription.unsubscribe();
+    subscription = null;
+    originalDelete();
+  };
+}
